microscopy cell profiler file renaming
This will attempt to rename files with xy in the title. Please type in
the number you wish the images titles to be renamed to. The program
finds files with xy in the filename and renumbers the xy location
starting from your input number. Currently set up for 3 channel
sequences